+++ src/bda/plone/orders/browser/views.py
@@ -1,4 +1,5 @@
 from AccessControl import Unauthorized
+from Products.CMFPlone.utils import getToolByName
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.statusmessages.interfaces import IStatusMessage
@@ -9,6 +10,7 @@ from bda.plone.checkout.vocabularies import get_pycountry_name
 from bda.plone.orders import interfaces as ifaces
 from bda.plone.orders import message_factory as _
 from bda.plone.orders import permissions
+from bda.plone.orders import safe_encode
 from bda.plone.orders import vocabularies as vocabs
 from bda.plone.orders.common import DT_FORMAT
 from bda.plone.orders.common import OrderData
@@ -19,11 +21,12 @@ from bda.plone.orders.common import get_orders_soup
 from bda.plone.orders.common import get_vendor_by_uid
 from bda.plone.orders.common import get_vendor_uids_for
 from bda.plone.orders.common import get_vendors_for
-from bda.plone.payment import Payments
+from bda.plone.shop.interfaces import IBuyable  # XXX: dependency inversion
 from decimal import Decimal
 from odict import odict
 from plone.app.uuid.utils import uuidToURL
 from plone.memoize import view
+from plone.uuid.interfaces import IUUID
 from repoze.catalog.query import Any
 from repoze.catalog.query import Contains
 from repoze.catalog.query import Eq
@@ -340,11 +343,15 @@ class OrdersTableBase(BrowserView):
         return None
 
     def render_salaried(self, colname, record):
-        return OrderData(self.context, order=record).salaried\
+        salaried = OrderData(self.context, order=record).salaried\
             or ifaces.SALARIED_NO
+        return translate(vocabs.salaried_vocab()[salaried],
+                         context=self.request)
 
     def render_state(self, colname, record):
-        return OrderData(self.context, order=record).state
+        state = OrderData(self.context, order=record).state
+        return translate(vocabs.state_vocab()[state],
+                         context=self.request)
 
     def render_dt(self, colname, record):
         value = record.attrs.get(colname, '')
@@ -508,12 +515,16 @@ class OrdersTable(OrdersTableBase):
 
     def render_salaried(self, colname, record):
         if not self.check_modify_order(record):
-            return OrderData(self.context, order=record).salaried
+            salaried = OrderData(self.context, order=record).salaried
+            return translate(vocabs.salaried_vocab()[salaried],
+                             context=self.request)
         return SalariedDropdown(self.context, self.request, record).render()
 
     def render_state(self, colname, record):
         if not self.check_modify_order(record):
-            return OrderData(self.context, order=record).state
+            state = OrderData(self.context, order=record).state
+            return translate(vocabs.state_vocab()[state],
+                             context=self.request)
         return StateDropdown(self.context, self.request, record).render()
 
     @property
@@ -574,7 +585,7 @@ class OrdersData(OrdersTable, TableData):
         if vendor_uid:
             vendor_uid = uuid.UUID(vendor_uid)
             # raise if given vendor uid not in user vendor uids
-            if not vendor_uid in vendor_uids:
+            if vendor_uid not in vendor_uids:
                 raise Unauthorized
             query = Any('vendor_uids', [vendor_uid])
         else:
@@ -651,7 +662,28 @@ class OrderViewBase(BrowserView):
         return ascur(self.order_data.discount_vat)
 
     @property
+    def shipping_title(self):
+        # XXX: either failure in upgrade step or node.ext.zodb bug
+        #      figure out
+        # order = self.order
+        order = self.order_data.order.attrs
+        title = translate(order['shipping_label'], context=self.request)
+        if order['shipping_description']:
+            title += ' (%s)' % translate(order['shipping_description'],
+                                         context=self.request)
+        return title
+
+    @property
+    def shipping_net(self):
+        return ascur(self.order_data.shipping_net)
+
+    @property
+    def shipping_vat(self):
+        return ascur(self.order_data.shipping_vat)
+
+    @property
     def shipping(self):
+        # B/C
         return ascur(self.order_data.shipping)
 
     @property
@@ -698,11 +730,7 @@ class OrderViewBase(BrowserView):
 
     @property
     def payment(self):
-        name = self.order['payment_selection.payment']
-        payment = Payments(self.context).get(name)
-        if payment:
-            return payment.label
-        return name
+        return self.order['payment_label']
 
     @property
     def salaried(self):
@@ -864,7 +892,7 @@ class DirectOrderView(OrderViewBase):
                 order = None
         if not email:
             err = _('anon_auth_err_email',
-                    default=u'Please provide the emailadress you used for '
+                    default=u'Please provide the email adress you used for '
                             u'submitting the order.')
             errs.append(err)
         if not ordernumber:
@@ -942,6 +970,19 @@ def resolve_buyable_url(context, booking):
 COMPUTED_BOOKING_EXPORT_ATTRS['url'] = resolve_buyable_url
 
 
+def cleanup_for_csv(value):
+    """Cleanup a value for CSV export.
+    """
+    if isinstance(value, datetime.datetime):
+        value = value.strftime(DT_FORMAT)
+    if value == '-':
+        value = ''
+    if isinstance(value, float) or \
+       isinstance(value, Decimal):
+        value = str(value).replace('.', ',')
+    return safe_encode(value)
+
+
 class ExportOrdersForm(YAMLForm):
     browser_template = ViewPageTemplateFile('export.pt')
     form_template = 'bda.plone.orders.browser:forms/orders_export.yaml'
@@ -986,15 +1027,11 @@ class ExportOrdersForm(YAMLForm):
         self.to_date = data.fetch('exportorders.to').extracted
 
     def export_val(self, record, attr_name):
+        """Get attribute from record and cleanup.
+        Since the record object is available, you can return aggregated values.
+        """
         val = record.attrs.get(attr_name)
-        if isinstance(val, datetime.datetime):
-            val = val.strftime(DT_FORMAT)
-        if val == '-':
-            val = ''
-        if isinstance(val, float) or \
-           isinstance(val, Decimal):
-            val = str(val).replace('.', ',')
-        return val
+        return cleanup_for_csv(val)
 
     def csv(self, request):
         # get orders soup
@@ -1010,7 +1047,7 @@ class ExportOrdersForm(YAMLForm):
         if vendor_uid:
             vendor_uid = uuid.UUID(vendor_uid)
             # raise if given vendor uid not in user vendor uids
-            if not vendor_uid in vendor_uids:
+            if vendor_uid not in vendor_uids:
                 raise Unauthorized
             query = query & Any('vendor_uids', [vendor_uid])
         else:
@@ -1042,6 +1079,7 @@ class ExportOrdersForm(YAMLForm):
             for attr_name in COMPUTED_ORDER_EXPORT_ATTRS:
                 cb = COMPUTED_ORDER_EXPORT_ATTRS[attr_name]
                 val = cb(self.context, order_data)
+                val = cleanup_for_csv(val)
                 order_attrs.append(val)
             for booking in order_data.bookings:
                 booking_attrs = list()
@@ -1053,6 +1091,7 @@ class ExportOrdersForm(YAMLForm):
                 for attr_name in COMPUTED_BOOKING_EXPORT_ATTRS:
                     cb = COMPUTED_BOOKING_EXPORT_ATTRS[attr_name]
                     val = cb(self.context, booking)
+                    val = cleanup_for_csv(val)
                     booking_attrs.append(val)
                 ex.writerow(order_attrs + booking_attrs)
                 booking.attrs['exported'] = True
@@ -1064,7 +1103,106 @@ class ExportOrdersForm(YAMLForm):
         self.request.response.setHeader('Content-Type', 'text/csv')
         self.request.response.setHeader('Content-Disposition',
                                         'attachment; filename=%s' % filename)
-        return sio.getvalue().decode('utf8')
+        ret = sio.getvalue()
+        sio.close()
+        return ret
+
+
+class ExportOrdersContextual(BrowserView):
+
+    def __call__(self):
+        user = plone.api.user.get_current()
+        # check if authenticated user is vendor
+        if not user.checkPermission(permissions.ModifyOrders, self.context):
+            raise Unauthorized
+
+        filename = '{}_{}.csv'.format(
+            safe_encode(self.context.title),
+            datetime.datetime.now().strftime('%Y-%m-%d_%H-%M'))
+        resp = self.request.response
+        resp.setHeader('content-type', 'text/csv; charset=utf-8')
+        resp.setHeader(
+            'content-disposition', 'attachment;filename={}'.format(filename))
+        return self.get_csv()
+
+    def export_val(self, record, attr_name):
+        """Get attribute from record and cleanup.
+        Since the record object is available, you can return aggregated values.
+        """
+        val = record.attrs.get(attr_name)
+        return cleanup_for_csv(val)
+
+    def get_csv(self):
+        context = self.context
+
+        # prepare csv writer
+        sio = StringIO()
+        ex = csv.writer(sio, dialect='excel-colon')
+        # exported column keys as first line
+        ex.writerow(ORDER_EXPORT_ATTRS +
+                    COMPUTED_ORDER_EXPORT_ATTRS.keys() +
+                    BOOKING_EXPORT_ATTRS +
+                    COMPUTED_BOOKING_EXPORT_ATTRS.keys())
+
+        bookings_soup = get_bookings_soup(context)
+
+        # First, filter by allowed vendor areas
+        vendor_uids = get_vendor_uids_for()
+        query_b = Any('vendor_uid', vendor_uids)
+
+        # Second, query for the buyable
+        query_cat = {}
+        query_cat['object_provides'] = IBuyable.__identifier__
+        query_cat['path'] = '/'.join(context.getPhysicalPath())
+        cat = getToolByName(context, 'portal_catalog')
+        res = cat(**query_cat)
+        buyable_uids = [IUUID(it.getObject()) for it in res]
+
+        query_b = query_b & Any('buyable_uid', buyable_uids)
+
+        all_orders = {}
+        for booking in bookings_soup.query(query_b):
+            booking_attrs = list()
+            # booking export attrs
+            for attr_name in BOOKING_EXPORT_ATTRS:
+                val = self.export_val(booking, attr_name)
+                booking_attrs.append(val)
+            # computed booking export attrs
+            for attr_name in COMPUTED_BOOKING_EXPORT_ATTRS:
+                cb = COMPUTED_BOOKING_EXPORT_ATTRS[attr_name]
+                val = cb(context, booking)
+                val = cleanup_for_csv(val)
+                booking_attrs.append(val)
+
+            # create order_attrs, if it doesn't exist
+            order_uid = booking.attrs.get('order_uid')
+            if order_uid not in all_orders:
+                order = get_order(context, order_uid)
+                order_data = OrderData(context,
+                                       order=order,
+                                       vendor_uids=vendor_uids)
+                order_attrs = list()
+                # order export attrs
+                for attr_name in ORDER_EXPORT_ATTRS:
+                    val = self.export_val(order, attr_name)
+                    order_attrs.append(val)
+                # computed order export attrs
+                for attr_name in COMPUTED_ORDER_EXPORT_ATTRS:
+                    cb = COMPUTED_ORDER_EXPORT_ATTRS[attr_name]
+                    val = cb(self.context, order_data)
+                    val = cleanup_for_csv(val)
+                    order_attrs.append(val)
+                all_orders[order_uid] = order_attrs
+
+            ex.writerow(all_orders[order_uid] + booking_attrs)
+
+            # TODO: also set for contextual exports? i'd say no.
+            # booking.attrs['exported'] = True
+            # bookings_soup.reindex(booking)
+
+        ret = sio.getvalue()
+        sio.close()
+        return ret
 
 
 class ReservationDone(BrowserView):