+++ src/bda/plone/orders/common.py
@@ -13,9 +13,10 @@ from bda.plone.checkout import CheckoutError
 from bda.plone.orders import permissions
 from bda.plone.orders import interfaces as ifaces
 from bda.plone.orders.interfaces import IVendor
+from bda.plone.payment import Payments
 from bda.plone.payment.interfaces import IPaymentData
 from bda.plone.shipping import Shippings
-from bda.plone.shop.interfaces import IBuyable
+from bda.plone.shop.interfaces import IBuyable # XXX: dependency inversion
 from decimal import Decimal
 from node.ext.zodb import OOBTNode
 from node.utils import instance_property
@@ -239,7 +240,7 @@ class OrderCheckoutAdapter(CheckoutAdapter):
 
     @property
     def skip_payment(self):
-        # TODO: seperate hirning session required
+        # XXX: separate braining session required
         order_data = OrderData(context=self.context, order=self.order)
         return SKIP_PAYMENT_IF_RESERVED \
             and order_data.state in (ifaces.STATE_RESERVED, ifaces.STATE_MIXED)
@@ -260,24 +261,49 @@ class OrderCheckoutAdapter(CheckoutAdapter):
 
     def save(self, providers, widget, data):
         super(OrderCheckoutAdapter, self).save(providers, widget, data)
+        order = self.order
+        # order UUID
+        uid = order.attrs['uid'] = uuid.uuid4()
+        # order creator
         creator = None
         member = plone.api.user.get_current()
         if member:
             creator = member.getId()
+        order.attrs['creator'] = creator
+        # order creation date
         created = datetime.datetime.now()
-        order = self.order
-        # XXX: payment name
-        #      payment type
-        #      payment total
-        #      shipping name
-        #      shipping vat
+        order.attrs['created'] = created
+        # payment related information
+        # XXX: handle no payment
+        pid = data.fetch('checkout.payment_selection.payment').extracted
+        payment = Payments(self.context).get(pid)
+        order.attrs['payment_method'] = pid
+        if payment:
+            order.attrs['payment_label'] = payment.label
+        else:
+            order.attrs['payment_label'] = _('unknown', default=u'Unknown')
+        # shipping related information
+        # XXX: handle no shipping
         sid = data.fetch('checkout.shipping_selection.shipping').extracted
         shipping = Shippings(self.context).get(sid)
-        order.attrs['shipping'] = shipping.calculate(self.items)
-        uid = order.attrs['uid'] = uuid.uuid4()
-        order.attrs['creator'] = creator
-        order.attrs['created'] = created
+        order.attrs['shipping_method'] = sid
+        order.attrs['shipping_label'] = shipping.label
+        order.attrs['shipping_description'] = shipping.description
+        try:
+            shipping_net = shipping.net(self.items)
+            shipping_vat = shipping.vat(self.items)
+            order.attrs['shipping_net'] = shipping_net
+            order.attrs['shipping_vat'] = shipping_vat
+            order.attrs['shipping'] = shipping_net + shipping_vat
+        # B/C for bda.plone.shipping < 0.4
+        except NotImplementedError:
+            shipping_net = shipping.calculate(self.items)
+            order.attrs['shipping_net'] = shipping_net
+            order.attrs['shipping_vat'] = Decimal(0)
+            order.attrs['shipping'] = shipping_net
+        # create order bookings
         bookings = self.create_bookings(order)
+        # lookup booking uids and vendor uids
         booking_uids = list()
         vendor_uids = set()
         for booking in bookings:
@@ -285,19 +311,24 @@ class OrderCheckoutAdapter(CheckoutAdapter):
             vendor_uids.add(booking.attrs['vendor_uid'])
         order.attrs['booking_uids'] = booking_uids
         order.attrs['vendor_uids'] = list(vendor_uids)
+        # cart discount related information
         cart_data = get_data_provider(self.context)
         cart_discount = cart_data.discount(self.items)
         order.attrs['cart_discount_net'] = cart_discount['net']
         order.attrs['cart_discount_vat'] = cart_discount['vat']
+        # create ordernumber
         orders_soup = get_orders_soup(self.context)
         ordernumber = create_ordernumber()
         while self.ordernumber_exists(orders_soup, ordernumber):
             ordernumber = create_ordernumber()
         order.attrs['ordernumber'] = ordernumber
+        # add order
         orders_soup.add(order)
+        # add bookings
         bookings_soup = get_bookings_soup(self.context)
         for booking in bookings:
             bookings_soup.add(booking)
+        # return uid of added order
         return uid
 
     def create_bookings(self, order):
@@ -475,6 +506,16 @@ class OrderData(object):
         return float(self.order.attrs['cart_discount_vat'])
 
     @property
+    def shipping_net(self):
+        # XXX: use decimal
+        return float(self.order.attrs['shipping_net'])
+
+    @property
+    def shipping_vat(self):
+        # XXX: use decimal
+        return float(self.order.attrs['shipping_vat'])
+
+    @property
     def shipping(self):
         # XXX: use decimal
         return float(self.order.attrs['shipping'])
@@ -520,7 +561,7 @@ class BuyableData(object):
         context = self.context
         bookings_soup = get_bookings_soup(context)
         order_bookings = dict()
-        for booking in bookings_soup.query(Eq('buyable_uid', context.UID())):
+        for booking in bookings_soup.query(Eq('buyable_uid', IUUID(context))):
             bookings = order_bookings.setdefault(
                 booking.attrs['order_uid'], list())
             bookings.append(booking)