+++ src/bda/plone/orders/mailtemplates.py
@@ -43,6 +43,8 @@ Ordered items:
 %(item_listing)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 RESERVATION_BODY_EN = """
@@ -74,6 +76,8 @@ Ordered items:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 DELIVERY_ADDRESS_EN = """
@@ -124,6 +128,8 @@ Bestellte Artikel:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 RESERVATION_BODY_DE = """
@@ -155,6 +161,8 @@ Bestellte Artikel:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 DELIVERY_ADDRESS_DE = """
@@ -205,6 +213,8 @@ Produit commandé:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 RESERVATION_BODY_FR = """
@@ -236,6 +246,8 @@ Produit commandé:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 DELIVERY_ADDRESS_FR = """
@@ -285,6 +297,8 @@ Articolo ordinato:
 Totale: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 RESERVATION_BODY_IT = """
@@ -315,6 +329,8 @@ Articolo ordinato:
 Totale: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 DELIVERY_ADDRESS_IT = """
@@ -364,6 +380,8 @@ Bestilte produkter:
 Total: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 RESERVATION_BODY_NO = """
@@ -395,6 +413,8 @@ Bestilte produkter:
 Totale: %(order_total)s
 
 %(global_text)s
+
+%(payment_text)s
 """
 
 DELIVERY_ADDRESS_NO = """
@@ -544,7 +564,6 @@ class DynamicMailTemplate(object):
         template
             a unicode string meant to be rendered using python string format
             method
-
         """
         assert isinstance(template, unicode), 'template must be unicode'
         try:
@@ -556,7 +575,8 @@ class DynamicMailTemplate(object):
         return True, ''
 
     def __call__(self, template, data):
-        """render template with data"""
+        """render template with data
+        """
         assert isinstance(template, unicode), 'template must be unicode'
         for key in self.required:
             if key not in data:
@@ -565,6 +585,7 @@ class DynamicMailTemplate(object):
 
 
 DYNAMIC_MAIL_LIBRARY_KEY = "bda.plone.order.dynamic_mail_lib"
+
 
 @implementer(IDynamicMailTemplateLibrary)
 class DynamicMailTemplateLibraryAquierer(object):
@@ -640,4 +661,3 @@ class DynamicMailTemplateLibraryStorage(DynamicMailTemplateLibraryAquierer):
 
     def __delitem__(self, name):
         del self._storage[name]
-