+++ src/bda/plone/orders/upgrades.py
@@ -1,7 +1,9 @@
+from bda.plone.orders import message_factory as _
 from bda.plone.orders.common import acquire_vendor_or_shop_root
 from bda.plone.orders.common import get_bookings_soup
 from bda.plone.orders.common import get_orders_soup
 from bda.plone.orders.common import OrderData
+from bda.plone.payment import Payments
 from decimal import Decimal
 from plone.app.uuid.utils import uuidToObject
 from plone.uuid.interfaces import IUUID
@@ -147,3 +149,74 @@ def fix_discount_attrs(ctx=None):
     if need_rebuild:
         bookings_soup.rebuild()
         logging.info("Rebuilt bookings catalog")
+
+
+def fix_shipping_attrs(ctx=None):
+    portal = getSite()
+    orders_soup = get_orders_soup(portal)
+    data = orders_soup.storage.data
+    for item in data.values():
+        if not 'shipping_method' in item.attrs:
+            item.attrs['shipping_method'] = 'unknown'
+            logging.info(
+                "Added shipping_method {0} to booking {1}".format(
+                    'unknown', item.attrs['uid']
+                )
+            )
+        if not 'shipping_label' in item.attrs:
+            item.attrs['shipping_label'] = _('unknown', default=u'Unknown')
+            logging.info(
+                "Added shipping_label {0} to booking {1}".format(
+                    'unknown', item.attrs['uid']
+                )
+            )
+        if not 'shipping_description' in item.attrs:
+            item.attrs['shipping_description'] = \
+                _('unknown', default=u'Unknown')
+            logging.info(
+                "Added shipping_description {0} to booking {1}".format(
+                    'unknown', item.attrs['uid']
+                )
+            )
+        if not 'shipping_net' in item.attrs:
+            item.attrs['shipping_net'] = item.attrs['shipping']
+            logging.info(
+                "Added shipping_net {0} to booking {1}".format(
+                    item.attrs['shipping'], item.attrs['uid']
+                )
+            )
+        if not 'shipping_vat' in item.attrs:
+            item.attrs['shipping_vat'] = Decimal(0)
+            logging.info(
+                "Added shipping_vat {0} to booking {1}".format(
+                    Decimal(0), item.attrs['uid']
+                )
+            )
+
+
+def fix_payment_attrs(ctx=None):
+    portal = getSite()
+    payments = Payments(portal)
+    orders_soup = get_orders_soup(portal)
+    data = orders_soup.storage.data
+    for item in data.values():
+        if not 'payment_method' in item.attrs:
+            continue
+        payment_method = item.attrs['payment_selection.payment']
+        payment = payments.get(payment_method)
+        if payment:
+            payment_label = payment.label
+        else:
+            payment_label = _('unknown', default=u'Unknown')
+        item.attrs['payment_method'] = payment_method
+        logging.info(
+            "Added payment_method {0} to booking {1}".format(
+                payment_method, item.attrs['uid']
+            )
+        )
+        item.attrs['payment_label'] = payment_label
+        logging.info(
+            "Added payment_label {0} to booking {1}".format(
+                payment_label, item.attrs['uid']
+            )
+        )